Time: 0 ms (100%), Space: 8.7 MB (36.14%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         if(x < 0) return false;
-        int dumX = x;
+        long int dumX = x;
         long int dummy = 0;
         while(x!=0){
             dummy = dummy*10 + x%10;
